Changed to doubles. Still need to resolve issue with making posts.

+++ Java/app/src/main/java/com/rao/domain/Asset.java
@@ -3,19 +3,19 @@ package com.rao.domain;
 public class Asset {
     private int category;
     private String lineItem;
-    private float amount;
+    private double amount;
 
-    public Asset(int category, String lineItem, float amount) {
+    public Asset(int category, String lineItem, double amount) {
         this.setCategory(category);
         this.setLineItem(lineItem);
         this.setAmount(amount);
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
