+++ Java/app/src/main/java/com/rao/service/NetWorthService.java
@@ -30,33 +30,33 @@ public class NetWorthService {
 
     public NetWorth initialize() {
         ArrayList<Asset> assets = new ArrayList<>();
-        assets.add(new Asset(1, "Chequing", 2000.00f));
-        assets.add(new Asset(1, "Savings for Taxes", 4000.00f));
-        assets.add(new Asset(1, "Rainy Day Fund", 506.00f));
-        assets.add(new Asset(1, "Savings for Fun", 5000.00f));
-        assets.add(new Asset(1, "Savings for Travel", 400.00f));
-        assets.add(new Asset(1, "Savings for Personal Development", 200.00f));
-        assets.add(new Asset(1, "Investment 1", 5000.00f));
-        assets.add(new Asset(1, "Investment 2", 60000.00f));
-        assets.add(new Asset(1, "Investment 3", 24000.00f));
-        assets.add(new Asset(2, "Primay Home", 455000.00f));
-        assets.add(new Asset(2, "Secondary Home", 1564321.00f));
-        assets.add(new Asset(2, "Other", 782.43f));
+        assets.add(new Asset(1, "Chequing", 2000.00));
+        assets.add(new Asset(1, "Savings for Taxes", 4000.00));
+        assets.add(new Asset(1, "Rainy Day Fund", 506.00));
+        assets.add(new Asset(1, "Savings for Fun", 5000.00));
+        assets.add(new Asset(1, "Savings for Travel", 400.00));
+        assets.add(new Asset(1, "Savings for Personal Development", 200.00));
+        assets.add(new Asset(1, "Investment 1", 5000.00));
+        assets.add(new Asset(1, "Investment 2", 60000.00));
+        assets.add(new Asset(1, "Investment 3", 24000.00));
+        assets.add(new Asset(2, "Primay Home", 455000.00));
+        assets.add(new Asset(2, "Secondary Home", 1564321.00));
+        assets.add(new Asset(2, "Other", 782.43));
         ArrayList<Liability> liabilities = new ArrayList<>();
-        liabilities.add(new Liability(1, "Credit Card 1", 200.00f, 4342.00f));
-        liabilities.add(new Liability(1, "Credit Card 2", 150.00f, 322.00f));
-        liabilities.add(new Liability(2, "Mortgage 1", 2000.00f, 250999.00f));
-        liabilities.add(new Liability(2, "Mortgage 2", 3500.00f, 632634.00f));
-        liabilities.add(new Liability(2, "Line of Credit", 500.00f, 10000.00f));
-        liabilities.add(new Liability(2, "Investment Loan", 700.00f, 12500.00f));
+        liabilities.add(new Liability(1, "Credit Card 1", 200.00, 4342.00));
+        liabilities.add(new Liability(1, "Credit Card 2", 150.00, 322.00));
+        liabilities.add(new Liability(2, "Mortgage 1", 2000.00, 250999.00));
+        liabilities.add(new Liability(2, "Mortgage 2", 3500.00, 632634.00));
+        liabilities.add(new Liability(2, "Line of Credit", 500.00, 10000.00));
+        liabilities.add(new Liability(2, "Investment Loan", 700.00, 12500.00));
         NetWorthCurrency netWorthCurrency = new NetWorthCurrency(Currency.CAD, "$");
         return calculateNetWorth(assets, liabilities, netWorthCurrency);
     }
 
     public NetWorth calculateNetWorth(ArrayList<Asset> assets, ArrayList<Liability> liabilities,
             NetWorthCurrency currency) {
-        float totalAssets = 0f;
-        float totalLiabilities = 0f;
+        double totalAssets = 0.00;
+        double totalLiabilities = 0.00;
         for (Asset asset : assets) {
             totalAssets += asset.getAmount();
         }
@@ -71,11 +71,11 @@ public class NetWorthService {
             throws CurrencyNotSupportedException, JSONException, StorageException, EndpointException, ServiceException {
         for (Asset asset : netWorth.getAssets()) {
             asset.setAmount(converter.convertCurrency(new BigDecimal(asset.getAmount()),
-                    netWorth.getCurrency().getCurrencyCode(), currency.getCurrencyCode()).floatValue());
+                    netWorth.getCurrency().getCurrencyCode(), currency.getCurrencyCode()).doubleValue());
         }
         for (Liability liability : netWorth.getLiabilities()) {
             liability.setAmount(converter.convertCurrency(new BigDecimal(liability.getAmount()),
-                    netWorth.getCurrency().getCurrencyCode(), currency.getCurrencyCode()).floatValue());
+                    netWorth.getCurrency().getCurrencyCode(), currency.getCurrencyCode()).doubleValue());
         }
         return calculateNetWorth(netWorth.getAssets(), netWorth.getLiabilities(), currency);
     }
